Changed Navigation to vertical for site but still needs work.
The navigation that I have planned out needs to be vertical so I used
the columns to make a side navigation using an <ul>. Positioning needs
to be fixed so that the links are centered and also that the navigation
area touches the bottom of the screen and is able to scroll with a
user. The inline styling of background-color on the HTML elements are
for test purposes only.