Added a way of finding all icons that are uploaded to WordPress with certain ending file name.
I need to add the href that each image is also suppose to be holding so
that they can correspond to their social media sites. Also I would like
to add another folder to WordPress to organize images, but the problem
with that is that there is no way of doing this without a plugin. This
functionality is what I will have to make as well.